requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
python-multipart==0.0.9
aiofiles==25.1.0
torch==2.9.0
torchvision==0.24.0
numpy==2.2.4
opencv-python==4.12.0.88
easyocr==1.7.2
scikit-image==0.25.2
google-generativeai==0.8.5
pandas==2.3.3
openpyxl==3.1.5
pillow==12.0.0
matplotlib==3.9.1
selenium==4.38.0
webdriver-manager==4.0.2
pyautogui==0.9.54
python-dotenv==1.0.1
certifi
gunicorn
